+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.0.0 (unreleased)
 ------------------
 
+- Added full hierarchy filter in containers listings
 - #15 Fixed imports to senaite.app.listing
 - Do not display Containers and SampleContainers in navbar by default
 - Compatibility with senaite.lims 2.x